requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
effdet==0.4.1
emoji==2.14.0
environs==11.2.1
et_xmlfile==2.0.0
eval_type_backport==0.2.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-vision==3.8.1
googleapis-common-protos==1.66.0
groq==0.13.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
iopath==0.1.10
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-groq==0.2.1
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langdetect==1.0.9
langgraph==0.2.53
langgraph-checkpoint==2.0.8
langgraph-sdk==0.1.40
langsmith==0.1.147
layoutparser==0.3.4
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.56.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.21.0
pikepdf==9.4.2
pillow==11.0.0
portalocker==3.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pypandoc==1.14
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.19
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
selenium==4.27.1
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.9
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.18.3
